Split streams module out of quinn::connection
